+++ projects/generic-trello/superman/api/routes/register.js
@@ -1,8 +1,8 @@
 import app from "./configs/app.js"
 import { Mongoose, UsersSchema } from './configs/mongo.js'
 import bcrypt from 'bcrypt'
-import { userSchema } from './configs/yupValidation.js'
-import { validateUser } from "./configs/validation.js"
+import { userSchema } from './configs/validation.js'
+import { validateUser } from "./configs/userValidation.js"
 
 
 app.post('/register', validateUser(userSchema), async (req, res) => {
@@ -18,12 +18,12 @@ app.post('/register', validateUser(userSchema), async (req, res) => {
     const foundUser = await Users.findOne({ email: userData.email })
     
     if(foundUser) {
-      return res.status(409).send({ msg: 'E-mail já cadastrado!' })
+      return res.status(409).send({ message: 'E-mail já cadastrado!' })
     }
       const user = new Users({ name: userData.name, email: userData.email, password: userData.password })
       await user.save()
 
-      res.status(201).send({ msg: 'Cadastro realizado com sucesso!' })
+      res.status(201).send({ message: 'Cadastro realizado com sucesso!' })
   } catch (err) {
     res.send(err)
   }
